requirementsVDMN.txt
argh==0.26.2
asn1crypto==0.24.0
backcall==0.1.0
backports.weakref==1.0rc1
bleach==1.5.0
category-encoders==2.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.0
graphviz==0.11
h5py==2.9.0
hiddenlayer==0.2
html5lib==0.9999999
idna==2.6
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.14.0
joblib==0.14.0
kazoo==2.6.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Logbook==1.4.3
lxml==4.3.3
Markdown==2.2.0
matplotlib==3.1.0
nltk==3.4.5
numpy==1.16.2
opencv-python==4.0.0.21
pandas==0.24.1
parso==0.5.0
pathtools==0.1.2
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
protobuf==3.7.0
ptyprocess==0.6.0
py==1.8.0
py4j==0.10.7
pycrypto==2.6.1
Pygments==2.4.2
pygobject==3.26.1
pyparsing==2.4.0
pyshark==0.4.2.3
pyspark==2.4.4
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-pcapng==0.2
pytz==2018.9
pyxdg==0.25
PyYAML==3.13
qhoptim==1.1.0
requests==2.21.0
ruamel.yaml==0.15.89
scapy==2.4.2
scikit-learn==0.22
scipy==1.2.1
SecretStorage==2.3.1
six==1.11.0
sklearn==0.0
statsmodels==0.10.2
tabulate==0.8.3
tensorboardX==1.8
tensorflow-gpu==1.2.0
Theano==1.0.4
torch==1.3.1
torchtext==0.3.1
torchvision==0.4.2
torchviz==0.0.1
tqdm==4.31.1
traitlets==4.3.2
ujson==1.35
unattended-upgrades==0.1
urllib3==1.24.1
vizdoom==1.1.8
watchdog==0.9.0
wcwidth==0.1.7
Werkzeug==0.14.1
zat==0.3.7
